move map to first page
